+++ src/yt_random.h
@@ -16,12 +16,12 @@ bool yt_random_next(struct yt_random *random, float *value,
     struct yt_error *error);
 bool yt_random_one_based_single(struct yt_random *random, float range,
     float *value, struct yt_error *error);
-bool yt_random_integer(struct yt_random *random, int range, int *value,
+bool yt_random_integer(struct yt_random *random, int range, uint16_t *value,
     struct yt_error *error);
 bool yt_random_nested_single(struct yt_random *random, float count,
     float *range, float *value, struct yt_error *error);
 bool yt_random_nested_integer(struct yt_random *random, int count, int range,
-    int *value, struct yt_error *error);
+    uint16_t *value, struct yt_error *error);
 bool yt_random_market_bases(struct yt_random *random, float bases[3],
     struct yt_error *error);
